codex: move mcp_edit tests
